Added IServiceCollection extension method for Dependency Injection.
Created a ReportBuilder class to help with the configuration the Repoter instances.

Made the EmailService class public, added abstraction to it's configuration, making it independent of the Reporter class.
Can be used on it's own (might move to different project).

Created a EmailServiceBuilder class to help with the configuration of EmailService instances.

Added default implementation of IReport interface.